feat: modules have > 1 entry point & directly get the initiator from the bundler

+++ src/interfaces/IModularBundler.sol
@@ -8,9 +8,8 @@ pragma solidity >=0.5.0;
 interface IModularBundler {
     /// @notice Receives a call from the Morpho Bundler.
 
-    /// @notice Calls `module`, passing along `data` and the current `initiator`.
-    /// @notice Sends `value` native tokens to `module`.
-    function callModule(address module, bytes calldata data, uint value) external payable;
+    /// @notice Calls `module`, passing along `data` and `value` native tokens to `module`.
+    function callModule(address module, bytes calldata data, uint256 value) external payable;
 
     /// @notice Responds to calls from the current module.
     /// @dev Triggers `_multicall` logic during a callback.
